+++ unstructured/embed/__init__.py
@@ -4,6 +4,7 @@ from unstructured.embed.octoai import OctoAIEmbeddingEncoder
 from unstructured.embed.openai import OpenAIEmbeddingEncoder
 from unstructured.embed.vertexai import VertexAIEmbeddingEncoder
 from unstructured.embed.voyageai import VoyageAIEmbeddingEncoder
+from unstructured.embed.mixedbreadai import MixedbreadAIEmbeddingEncoder
 
 EMBEDDING_PROVIDER_TO_CLASS_MAP = {
     "langchain-openai": OpenAIEmbeddingEncoder,
@@ -11,5 +12,6 @@ EMBEDDING_PROVIDER_TO_CLASS_MAP = {
     "langchain-aws-bedrock": BedrockEmbeddingEncoder,
     "langchain-vertexai": VertexAIEmbeddingEncoder,
     "langchain-voyageai": VoyageAIEmbeddingEncoder,
+    "mixedbread-ai": MixedbreadAIEmbeddingEncoder,
     "octoai": OctoAIEmbeddingEncoder,
 }